+++ Wire/Wire.csproj
@@ -52,6 +52,7 @@
     <Compile Include="SerializerOptions.cs" />
     <Compile Include="StreamExtensions.cs" />
     <Compile Include="Surrogate.cs" />
+    <Compile Include="Utils.cs" />
     <Compile Include="ValueSerializers\ArraySerializer.cs" />
     <Compile Include="ValueSerializers\BoolSerializer.cs" />
     <Compile Include="ValueSerializers\ByteArraySerializer.cs" />
